+++ level_3/fss_extended_read/c/main.c
@@ -1,6 +1,6 @@
 #include <level_3/fss_extended_read.h>
 
-int main(const int argc, const f_string *argv) {
+int main(const unsigned long argc, const f_string *argv) {
   const f_console_arguments arguments = { argc, argv };
   fss_extended_read_data data = fss_extended_read_data_initialize;
